Moved all android dependencies into android specific package.
