+++ app/src/main/java/dnsfilter/android/dnsserverconfig/DNSServerConfigActivity.java
@@ -1,4 +1,4 @@
-package dnsfilter.dnsserverconfig;
+package dnsfilter.android.dnsserverconfig;
 
 import android.app.ActionBar;
 import android.app.Activity;
