+++ app/src/main/java/dnsfilter/android/dnsserverconfig/widget/DNSServerConfigEntryTestState.java
@@ -1,4 +1,4 @@
-package dnsfilter.dnsserverconfig.widget;
+package dnsfilter.android.dnsserverconfig.widget;
 
 public enum DNSServerConfigEntryTestState {
     FAIL,
