+++ app/src/main/java/dnsfilter/android/dnsserverconfig/widget/DNSServerConfigEntryValidationResult.java
@@ -1,4 +1,4 @@
-package dnsfilter.dnsserverconfig.widget;
+package dnsfilter.android.dnsserverconfig.widget;
 
 import java.util.Objects;
