+++ app/src/main/java/dnsfilter/android/dnsserverconfig/widget/NotDeserializableException.java
@@ -1,4 +1,4 @@
-package dnsfilter.dnsserverconfig.widget;
+package dnsfilter.android.dnsserverconfig.widget;
 
 public class NotDeserializableException extends Exception{
