+++ app/src/test/java/dnsfilter/android/tests/DNSServerConfigEntrySerializerTest.java
@@ -6,9 +6,9 @@ import org.junit.Test;
 import java.util.HashMap;
 import java.util.Map;
 
-import dnsfilter.dnsserverconfig.widget.DNSServerConfigEntry;
-import dnsfilter.dnsserverconfig.widget.DNSServerConfigEntrySerializer;
-import dnsfilter.dnsserverconfig.widget.DNSType;
+import dnsfilter.android.dnsserverconfig.widget.DNSServerConfigEntry;
+import dnsfilter.android.dnsserverconfig.widget.DNSServerConfigEntrySerializer;
+import dnsfilter.android.dnsserverconfig.widget.DNSType;
 
 public class DNSServerConfigEntrySerializerTest {
